pyproject.toml
[project]
name = "pytest-tap"
version = "3.5"
description = "Test Anything Protocol (TAP) reporting plugin for pytest"
readme = "docs/releases.rst"
authors = [
    { name = "Matt Layman", email = "matthewlayman@gmail.com" }
]
license = { text = "BSD" }
homepage = "https://github.com/python-tap/pytest-tap"
dependencies = [
    "pytest>=3.0",
    "tap.py>=3.2,<4.0"
]
requires-python = ">=3.9"
keywords = ["TAP", "unittest", "pytest"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing"
]

[project.entry-points.pytest11]
tap = "pytest_tap.plugin"

[dependency-groups]
dev = [
    "pytest-tap",
    "pytest-subtests",
    "tox>=4.24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/docs",
    "/tests",
]
sources = ["src"]

[tool.hatch.build.targets.wheel]
sources = ["src"]
packages = ["pytest_tap"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
markers = ["subtests"]

[tool.uv.sources]
pytest-tap = { workspace = true }

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # bandit: Use of `assert` detected
    "S101",
]
